scripts/requirements.txt
autopep8
veryfi
python-dotenv